requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
SQLAlchemy>=2.0
psycopg[binary]==3.1.18
pydantic>=2.7.4,<3
pydantic-settings==2.1.0
python-dotenv
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.10.0
anthropic==0.18.1
redis==5.0.1
celery==5.3.6
alembic>=1.13
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.2.2
numpy==2.1.2
python-dateutil==2.8.2
httpx==0.26.0
websockets==12.0
pytest==7.4.4
pytest-asyncio==0.23.3
email-validator==2.0.0
pypdf==4.*
google-api-python-client>=2.100.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
cryptography>=42
python-docx>=1.1.0
# parsing
pdfminer.six>=20231228
python-docx>=1.1.0
# (по желанию для RTF)
striprtf>=0.0.26
